Rename composites.py to assembler.py
Makes more sense for the default assembler to be an assembler.
