+++ tests/config/basic/storageClasses.yaml
@@ -15,7 +15,7 @@ storageClasses:
   StructuredData: &StructuredData
     # Data from a simple Python class
     pytype: examplePythonTypes.MetricsExample
-    assembler: lsst.daf.butler.core.composites.CompositeAssemblerMonolithic
+    assembler: lsst.daf.butler.CompositeAssemblerMonolithic
     # Use YAML formatter by default
     components:
       # Components are those supported by get.
@@ -28,7 +28,7 @@ storageClasses:
     <<: *StructuredData
   StructuredComposite: &StructuredComposite
     pytype: examplePythonTypes.MetricsExample
-    assembler: lsst.daf.butler.core.composites.CompositeAssembler
+    assembler: lsst.daf.butler.CompositeAssembler
     components:
       summary: StructuredDataDictYaml
       output: StructuredDataDictYaml
